comprehensionrequirements.txt
streamlit
opencv-python
mediapipe
numpy
pandas
Pillow
matplotlib
